refactor CTuneParam and CTuningParams out of the gamecore
