+++ src/game/gamecore.cpp
@@ -5,45 +5,7 @@
 
 #include <engine/shared/config.h>
 
-const char *CTuningParams::ms_apNames[] =
-	{
-#define MACRO_TUNING_PARAM(Name, ScriptName, Value, Description) #ScriptName,
-#include "tuning.h"
-#undef MACRO_TUNING_PARAM
-};
 
-bool CTuningParams::Set(int Index, float Value)
-{
-	if(Index < 0 || Index >= Num())
-		return false;
-	((CTuneParam *)this)[Index] = Value;
-	return true;
-}
-
-bool CTuningParams::Get(int Index, float *pValue) const
-{
-	if(Index < 0 || Index >= Num())
-		return false;
-	*pValue = (float)((CTuneParam *)this)[Index];
-	return true;
-}
-
-bool CTuningParams::Set(const char *pName, float Value)
-{
-	for(int i = 0; i < Num(); i++)
-		if(str_comp_nocase(pName, ms_apNames[i]) == 0)
-			return Set(i, Value);
-	return false;
-}
-
-bool CTuningParams::Get(const char *pName, float *pValue) const
-{
-	for(int i = 0; i < Num(); i++)
-		if(str_comp_nocase(pName, ms_apNames[i]) == 0)
-			return Get(i, pValue);
-
-	return false;
-}
 
 float HermiteBasis1(float v)
 {
